src/database/connection.py
import mysql.connector
from mysql.connector import Error

def create_connection():
    try:
        connection = mysql.connector.connect(
            host="127.0.0.1",
            user="root",
            password="",
            database="correo_de_yury"
        )
        if connection.is_connected():
            print("Conexión exitosa a la base de datos")
            return connection
    except Error as e:
        print(f"Error al conectar: {e}")
        return None

# Para ejecutar y probar la conexión:
conn = create_connection()
if conn:
    conn.close()