src/requirements.txt
mysql-connect-python
sqlalchemy
flask
jwt
